Functional simulation with centralized architecture complete! Now need to check how to reduce line count in SV.

+++ asic/func_sim/include/Compute_Verification.hpp
@@ -78,11 +78,11 @@ void verify_result(RESULT_TYPE type, float result, float* input_data, uint16_t s
 void verify_layer_out(COMPUTE_VERIFICATION_STEP cim_state, uint8_t id, float* data, uint16_t starting_addr, DATA_WIDTH data_width);
 #elif CENTRALIZED_ARCH
 bool are_equal(float a, float b, int32_t index);
-void verify_result(RESULT_TYPE type, float result, float* input_data, uint16_t starting_addr, uint16_t len, DATA_WIDTH data_width);
-void verify_layer_out(COMPUTE_VERIFICATION_STEP cim_state, float* data, uint16_t starting_addr, uint16_t outside_dim_len, DATA_WIDTH data_width);
+void verify_result(RESULT_TYPE type, float result, float* input_data, uint32_t starting_addr, uint16_t len, DATA_WIDTH data_width);
+void verify_layer_out(COMPUTE_VERIFICATION_STEP cim_state, float* data, uint32_t starting_addr, uint16_t outside_dim_len, DATA_WIDTH data_width);
 #endif
-void print_softmax_error(float* data, uint16_t starting_addr, DATA_WIDTH data_width);
-void verify_softmax_storage(float* intermediate_res, uint16_t prev_softmax_base_addr);
+void print_softmax_error(float* data, uint32_t starting_addr, DATA_WIDTH data_width);
+void verify_softmax_storage(float* intermediate_res, uint32_t prev_softmax_base_addr);
 void print_intermediate_value_stats();
 void reset_stats();
